+++ docs/tutorial.py
@@ -1,7 +1,6 @@
 # # Tutorial
 #
 # This tutorial provides a full worked example of using `ome-zarr-models`
-
 import os
 import tempfile
 
@@ -9,22 +8,22 @@ import matplotlib.pyplot as plt
 import numpy as np
 import zarr
 import zarr.storage
-from pydantic_zarr.v2 import ArraySpec
+from pydantic_zarr.v3 import AnyArraySpec, ArraySpec, NamedConfig
 from rich.pretty import pprint
 
 from ome_zarr_models import open_ome_zarr
-from ome_zarr_models.v04.axes import Axis
-from ome_zarr_models.v04.image import Image
+from ome_zarr_models.v05.axes import Axis
+from ome_zarr_models.v05.image import Image
 
 # ## Loading datasets
 #
 # OME-Zarr datasets are Zarr groups with specific metadata.
 # To open an OME-Zarr dataset, we first open the Zarr group.
 
-zarr_group = zarr.open(
-    "https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0062A/6001240.zarr", mode="r"
+zarr_group = zarr.open_group(
+    "https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.5/idr0066/ExpD_chicken_embryo_MIP.ome.zarr",
+    mode="r",
 )
-
 # If you're not sure what type or OME-Zarr version of data you have, you can
 # use `open_ome_zarr()` to automatically 'guess' the correct group:
 
@@ -41,15 +40,15 @@ ome_zarr_image = Image.from_zarr(zarr_group)
 # No errors, which means the metadata is valid 🎉
 #
 # ## Accessing metadata
-# To access the OME-Zarr metadata, use the `.attributes` property:
+# To access the OME-Zarr metadata, use the `.ome_attributes` property:
 
-metadata = ome_zarr_image.attributes
+metadata = ome_zarr_image.ome_attributes
 pprint(metadata)
 
 # And as an example of getting more specific metadata, lets get the metadata
 # for all the datasets in this multiscales:
 
-pprint(metadata.multiscales[0].datasets)
+pprint(ome_zarr_image.datasets[0])
 
 # ## Accessing data
 #
@@ -57,13 +56,12 @@ pprint(metadata.multiscales[0].datasets)
 # the Zarr arrays using the `zarr-python` library.
 # For example, to get the highest resolution image:
 
-zarr_arr = zarr_group[metadata.multiscales[0].datasets[0].path]
+zarr_arr = zarr_group[ome_zarr_image.datasets[0][3].path]
 pprint(zarr_arr)
 
-# To finish off this section on accessing data, lets plot the first z-slice of the
-# first channel of this data:
+# To finish off this section on accessing data, lets plot this image:
 
-plt.imshow(zarr_arr[0, 0, :, :], cmap="gray")
+plt.imshow(zarr_arr, cmap="gray")
 
 # ## Creating new datasets
 #
@@ -77,9 +75,36 @@ plt.imshow(zarr_arr[0, 0, :, :], cmap="gray")
 # First, we need to create `ArraySpec` objects, which tell `ome-zarr-models`
 # what the structure of the data arrays will be.
 
-array_specs = [
-    ArraySpec(shape=(100, 100), chunks=(32, 32), dtype=np.uint16),
-    ArraySpec(shape=(50, 50), chunks=(32, 32), dtype=np.uint16),
+
+array_specs: list[AnyArraySpec] = [
+    ArraySpec(
+        shape=(100, 100),
+        data_type=np.uint16,
+        chunk_grid=NamedConfig(
+            name="regular",
+            configuration={"chunk_shape": [32, 32]},
+        ),
+        chunk_key_encoding=NamedConfig(
+            name="default", configuration={"separator": "/"}
+        ),
+        fill_value=0,
+        codecs=[NamedConfig(name="bytes")],
+        dimension_names=["y", "x"],
+    ),
+    ArraySpec(
+        shape=(100, 100),
+        data_type=np.uint16,
+        chunk_grid=NamedConfig(
+            name="regular",
+            configuration={"chunk_shape": [32, 32]},
+        ),
+        chunk_key_encoding=NamedConfig(
+            name="default", configuration={"separator": "/"}
+        ),
+        fill_value=0,
+        codecs=[NamedConfig(name="bytes")],
+        dimension_names=["y", "x"],
+    ),
 ]
 
 # Next, we'll set some metadata values
@@ -99,7 +124,7 @@ ome_zarr_image = Image.new(
     scales=[[p * 1 for p in pixel_size], [p * 2 for p in pixel_size]],
     translations=[[0, 0], [p * 0.5 for p in pixel_size]],
 )
-print(ome_zarr_image)
+pprint(ome_zarr_image)
 
 # It's also possible to create array metadata from existing arrays.
 #
@@ -111,9 +136,9 @@ array_specs = [ArraySpec.from_array(arr0), ArraySpec.from_array(arr1)]
 
 # or for Zarr arrays:
 
-arr0 = zarr.zeros(shape=(100, 100), dtype=np.uint16)
-arr1 = zarr.zeros(shape=(50, 50), dtype=np.uint16)
-array_specs = [ArraySpec.from_array(arr0), ArraySpec.from_array(arr1)]
+arr_zarr0 = zarr.zeros(shape=(100, 100), dtype=np.uint16, zarr_format=2)
+arr_zarr1 = zarr.zeros(shape=(50, 50), dtype=np.uint16, zarr_format=2)
+array_specs = [ArraySpec.from_array(arr_zarr0), ArraySpec.from_array(arr_zarr1)]
 
 # ## Saving datasets
 #
@@ -127,6 +152,6 @@ array_specs = [ArraySpec.from_array(arr0), ArraySpec.from_array(arr1)]
 # to, and then list the directory to show that it has been saved.
 
 with tempfile.TemporaryDirectory() as fp:
-    store = zarr.DirectoryStore(path=fp)
+    store = zarr.storage.LocalStore(fp)
     ome_zarr_image.to_zarr(store=store, path="/")
     print(os.listdir(fp))