+++ src/ome_zarr_models/common/axes.py
@@ -14,12 +14,14 @@ AxisType = Literal["space", "time", "channel"]
 class Axis(BaseAttrs):
     """
     Model for an element of `Multiscale.axes`.
-
-    See https://ngff.openmicroscopy.org/0.4/#axes-md.
     """
 
-    name: str
+    # Explicitly name could be any JsonValue, but implicitly it must match Zarr array
+    # dimension_names which limits it to str | None
+
+    name: str | None = None
     type: str | None = None
+    # Unit probably intended to be str, but the spec doesn't explicitly specify
     unit: str | JsonValue | None = None