+++ src/ome_zarr_models/v04/image_label.py
@@ -1,10 +1,4 @@
-"""
-For reference, see the [image label section of the OME-Zarr specification](https://ngff.openmicroscopy.org/0.4/index.html#label-md).
-"""
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Self
+from typing import Self
 
 import zarr
 from pydantic import Field
@@ -15,13 +9,7 @@ from ome_zarr_models.v04.image import Image
 from ome_zarr_models.v04.image_label_types import Label
 from ome_zarr_models.v04.multiscales import Multiscale
 
-if TYPE_CHECKING:
-    import zarr
-
-__all__ = [
-    "ImageLabel",
-    "ImageLabelAttrs",
-]
+__all__ = ["ImageLabel", "ImageLabelAttrs"]
 
 
 class ImageLabelAttrs(BaseAttrs):
@@ -39,7 +27,7 @@ class ImageLabel(BaseGroupv04[ImageLabelAttrs]):
     """
 
     @classmethod
-    def from_zarr(cls, group: zarr.Group) -> Self:
+    def from_zarr(cls, group: zarr.Group) -> Self:  # type: ignore[override]
         """
         Create an instance of an OME-Zarr image from a `zarr.Group`.
 
@@ -50,4 +38,4 @@ class ImageLabel(BaseGroupv04[ImageLabelAttrs]):
         """
         # Use Image.from_zarr() to validate multiscale metadata
         Image.from_zarr(group)
-        return super().from_zarr(group)  # type: ignore[no-any-return]
+        return super().from_zarr(group)