+++ tests/v05/test_image.py
@@ -1,13 +1,39 @@
-from ome_zarr_models._v05.axes import Axis
-from ome_zarr_models._v05.coordinate_transformations import VectorScale
-from ome_zarr_models._v05.image import Image, ImageAttrs
-from ome_zarr_models._v05.multiscales import Dataset, Multiscale
-from tests.v05.conftest import json_to_zarr_group
+import re
+
+import pytest
+from pydantic import ValidationError
+from zarr.abc.store import Store
+
+from ome_zarr_models.v05.axes import Axis
+from ome_zarr_models.v05.coordinate_transformations import VectorScale
+from ome_zarr_models.v05.image import Image, ImageAttrs
+from ome_zarr_models.v05.labels import LabelsAttrs
+from ome_zarr_models.v05.multiscales import Dataset, Multiscale
+from tests.v05.conftest import json_to_dict, json_to_zarr_group
 
 
-def test_image() -> None:
-    zarr_group = json_to_zarr_group(json_fname="image_example.json")
+def test_image(store: Store) -> None:
+    zarr_group = json_to_zarr_group(json_fname="image_example.json", store=store)
+    zarr_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "1",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "2",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
     ome_group = Image.from_zarr(zarr_group)
+
     assert ome_group.attributes.ome == ImageAttrs(
         multiscales=[
             Multiscale(
@@ -56,3 +82,164 @@ def test_image() -> None:
         ],
         version="0.5",
     )
+
+
+def test_image_no_dim_names(store: Store) -> None:
+    zarr_group = json_to_zarr_group(json_fname="image_example.json", store=store)
+    zarr_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    arr1 = zarr_group.create_array("1", shape=(1, 1, 1, 1, 1), dtype="uint8")
+    assert arr1.metadata.dimension_names is None  # type: ignore[union-attr]
+    zarr_group.create_array(
+        "2",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    with pytest.raises(
+        ValidationError,
+        match="The array in this group at  '1' has no dimension_names metadata",
+    ):
+        Image.from_zarr(zarr_group)
+
+
+def test_image_wrong_dim_names(store: Store) -> None:
+    zarr_group = json_to_zarr_group(json_fname="image_example.json", store=store)
+    zarr_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "1",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "x", "y"],
+    )
+    zarr_group.create_array(
+        "2",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    with pytest.raises(
+        ValidationError,
+        match=re.escape(
+            "The multiscale metadata has ('t', 'c', 'z', 'y', 'x') axes names "
+            "which does not match the dimension names of the array "
+            "found in this group at path '1' (('t', 'c', 'z', 'x', 'y'))"
+        ),
+    ):
+        Image.from_zarr(zarr_group)
+
+
+def test_image_with_labels(store: Store) -> None:
+    zarr_group = json_to_zarr_group(json_fname="image_example.json", store=store)
+    zarr_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "1",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "2",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    # Add labels group
+    labels_group = zarr_group.create_group(
+        "labels",
+        attributes=json_to_dict(json_fname="labels_example.json"),
+    )
+
+    with pytest.raises(
+        ValueError, match="Label path 'cell_space_segmentation' not found in zarr group"
+    ):
+        Image.from_zarr(zarr_group)
+
+    # Add image labels group
+    image_label_group = labels_group.create_group(
+        "cell_space_segmentation",
+        attributes=json_to_dict(json_fname="image_label_example.json"),
+    )
+    image_label_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    image_label_group.create_array(
+        "1",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    image_label_group.create_array(
+        "2",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    image = Image.from_zarr(zarr_group)
+    assert image.labels is not None
+    assert image.labels.attributes.ome == LabelsAttrs(
+        version="0.5", labels=["cell_space_segmentation"]
+    )
+
+
+def test_image_with_labels_mismatch_multiscales(store: Store) -> None:
+    zarr_group = json_to_zarr_group(json_fname="image_example.json", store=store)
+    zarr_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "1",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    zarr_group.create_array(
+        "2",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    # Add labels group
+    labels_group = zarr_group.create_group(
+        "labels",
+        attributes=json_to_dict(json_fname="labels_example.json"),
+    )
+    # Add image labels group
+    image_label_group = labels_group.create_group(
+        "cell_space_segmentation",
+        attributes=json_to_dict(json_fname="labels_image_example.json"),
+    )
+    image_label_group.create_array(
+        "0",
+        shape=(1, 1, 1, 1, 1),
+        dtype="uint8",
+        dimension_names=["t", "c", "z", "y", "x"],
+    )
+    with pytest.raises(
+        RuntimeError,
+        match=re.escape(
+            "Number of image label multiscale levels (1) doesn't match "
+            "number of image multiscale levels (3)."
+        ),
+    ):
+        Image.from_zarr(zarr_group)